+++ packages/eslint-plugin-fiori-tools/src/language/diagnostics.ts
@@ -11,6 +11,8 @@ export const CREATION_MODE_FOR_TABLE = 'sap-creation-mode-for-table';
 export const STATE_PRESERVATION_MODE = 'sap-state-preservation-mode';
 export const TABLE_PERSONALIZATION = 'sap-table-personalization';
 export const TABLE_COLUMN_VERTICAL_ALIGNMENT = 'sap-table-column-vertical-alignment';
+export const CONDENSED_TABLE_LAYOUT = 'sap-condensed-table-layout';
+export const STRICT_UOM_FILTERING = 'sap-strict-uom-filtering';
 
 export interface WidthIncludingColumnHeaderDiagnostic {
     type: typeof WIDTH_INCLUDING_COLUMN_HEADER_RULE_TYPE;
@@ -110,6 +112,16 @@ export interface TableColumnVerticalAlignment {
     manifest: ManifestPropertyDiagnosticData;
 }
 
+export interface CondensedTableLayout {
+    type: typeof CONDENSED_TABLE_LAYOUT;
+    pageName: string;
+    manifest: ManifestPropertyDiagnosticData;
+}
+export interface StrictUomFiltering {
+    type: typeof STRICT_UOM_FILTERING;
+    manifest: ManifestPropertyDiagnosticData;
+}
+
 export type Diagnostic =
     | WidthIncludingColumnHeaderDiagnostic
     | AnchorBarVisible
@@ -119,5 +131,7 @@ export type Diagnostic =
     | EnableExport
     | EnablePaste
     | StatePreservationMode
+    | CondensedTableLayout
+    | StrictUomFiltering
     | TablePersonalization
     | TableColumnVerticalAlignment;