+++ packages/eslint-plugin-fiori-tools/src/project-context/linker/fe-v4.ts
@@ -7,6 +7,7 @@ import { collectTables, collectSections } from './annotations';
 
 export interface ApplicationSetting {
     createMode: string;
+    disableStrictUomFiltering: boolean;
 }
 
 export interface LinkedFeV4App extends ConfigurationBase<'fe-v4', ApplicationSetting> {
@@ -55,6 +56,7 @@ export interface TableSettings {
     disableCopyToClipboard: boolean;
     enableExport: boolean;
     enablePaste: boolean;
+    condensedTableLayout: boolean;
     personalization: boolean | { column?: boolean; filter?: boolean; sort?: boolean; group?: boolean };
 }
 
@@ -66,6 +68,9 @@ interface ManifestApplicationSettings {
         table?: {
             defaultCreationMode?: string;
         };
+    };
+    app?: {
+        disableStrictUomFiltering?: boolean;
     };
 }
 
@@ -139,6 +144,18 @@ function createTable(configurationKey: string, pathToPage: string[], table?: Tab
                     configurationKey,
                     'tableSettings',
                     'enablePaste'
+                ],
+                values: [true, false]
+            },
+            condensedTableLayout: {
+                configurationPath: [
+                    ...pathToPage,
+                    'options',
+                    'settings',
+                    'controlConfiguration',
+                    configurationKey,
+                    'tableSettings',
+                    'condensedTableLayout'
                 ],
                 values: [true, false]
             },
@@ -288,6 +305,7 @@ interface TableConfiguration {
         disableCopyToClipboard?: boolean;
         enableExport?: boolean;
         enablePaste?: boolean;
+        condensedTableLayout?: boolean;
         creationMode?: {
             name?: string;
         };
@@ -385,6 +403,8 @@ function linkListReportTable(
                 tableControl.configuration.enableExport.valueInFile = enableExportValue;
                 const enablePasteValue = controlConfiguration.tableSettings?.enablePaste;
                 tableControl.configuration.enablePaste.valueInFile = enablePasteValue;
+                const condensedTableLayoutValue = controlConfiguration.tableSettings?.condensedTableLayout;
+                tableControl.configuration.condensedTableLayout.valueInFile = condensedTableLayoutValue;
                 const creationModeValue = controlConfiguration.tableSettings?.creationMode?.name;
                 tableControl.configuration.creationMode.valueInFile = creationModeValue;
                 tableControl.configuration.creationMode.values = getCreationModeValues(tableType);
@@ -480,6 +500,8 @@ function linkObjectPageSections(
             tableControl.configuration.enableExport.valueInFile = enableExportValue;
             const enablePasteValue = controlConfiguration.tableSettings?.enablePaste;
             tableControl.configuration.enablePaste.valueInFile = enablePasteValue;
+            const condensedTableLayoutValue = controlConfiguration.tableSettings?.condensedTableLayout;
+            tableControl.configuration.condensedTableLayout.valueInFile = condensedTableLayoutValue;
             const creationModeValue = controlConfiguration.tableSettings?.creationMode?.name;
             tableControl.configuration.creationMode.valueInFile = creationModeValue;
             tableControl.configuration.creationMode.values = getCreationModeValues(tableType);
@@ -613,6 +635,7 @@ function resolveNavigationProperties(root: MetadataElement, segments: string[]):
 function linkApplicationSettings(context: LinkerContext): LinkedFeV4App {
     const config: ManifestApplicationSettings = context.app.manifestObject['sap.fe'] ?? {};
     const createMode = config.macros?.table?.defaultCreationMode;
+    const disableStrictUomFiltering = config.app?.disableStrictUomFiltering;
     const linkedApp: LinkedFeV4App = {
         type: 'fe-v4',
         pages: [],
@@ -621,6 +644,11 @@ function linkApplicationSettings(context: LinkerContext): LinkedFeV4App {
                 values: ['InlineCreationRows', 'NewPage'],
                 configurationPath: ['sap.fe', 'macros', 'table', 'defaultCreationMode'],
                 valueInFile: createMode
+            },
+            disableStrictUomFiltering: {
+                values: [true, false],
+                configurationPath: ['sap.fe', 'app', 'disableStrictUomFiltering'],
+                valueInFile: disableStrictUomFiltering
             }
         }
     };