+++ packages/project-integrity/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 0.2.61
 
+*Released: 2026-03-20T16:07:49Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.14
 
 ## 0.2.60
+
+*Released: 2026-03-05T16:08:22Z*
 
 ### Patch Changes
 
@@ -15,6 +19,8 @@
 
 ## 0.2.59
 
+*Released: 2026-03-05T12:30:25Z*
+
 ### Patch Changes
 
 - Updated dependencies [7c06ef0]
@@ -22,11 +28,15 @@
 
 ## 0.2.58
 
+*Released: 2026-03-04T14:39:10Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.11
 
 ## 0.2.57
+
+*Released: 2026-02-27T15:42:39Z*
 
 ### Patch Changes
 
@@ -35,11 +45,15 @@
 
 ## 0.2.56
 
+*Released: 2026-02-23T22:35:31Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.9
 
 ## 0.2.55
+
+*Released: 2026-02-23T19:07:48Z*
 
 ### Patch Changes
 
@@ -48,11 +62,15 @@
 
 ## 0.2.54
 
+*Released: 2026-02-20T21:31:31Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.7
 
 ## 0.2.53
+
+*Released: 2026-02-20T13:46:55Z*
 
 ### Patch Changes
 
@@ -62,6 +80,8 @@
 
 ## 0.2.52
 
+*Released: 2026-02-18T21:45:35Z*
+
 ### Patch Changes
 
 - Updated dependencies [d57cc47]
@@ -69,11 +89,15 @@
 
 ## 0.2.51
 
+*Released: 2026-02-16T16:08:11Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.4
 
 ## 0.2.50
+
+*Released: 2026-02-13T09:54:38Z*
 
 ### Patch Changes
 
@@ -82,6 +106,8 @@
 
 ## 0.2.49
 
+*Released: 2026-02-12T10:29:41Z*
+
 ### Patch Changes
 
 - Updated dependencies [d2b772d]
@@ -89,11 +115,15 @@
 
 ## 0.2.48
 
+*Released: 2026-02-10T23:50:15Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.35.1
 
 ## 0.2.47
+
+*Released: 2026-02-09T11:21:43Z*
 
 ### Patch Changes
 
@@ -102,12 +132,16 @@
 
 ## 0.2.46
 
+*Released: 2026-02-05T20:09:45Z*
+
 ### Patch Changes
 
 - Updated dependencies [467e6aa]
     - @sap-ux/project-access@1.34.7
 
 ## 0.2.45
+
+*Released: 2026-02-05T08:38:39Z*
 
 ### Patch Changes
 
@@ -116,6 +150,8 @@
 
 ## 0.2.44
 
+*Released: 2026-02-03T15:42:23Z*
+
 ### Patch Changes
 
 - Updated dependencies [89175fe]
@@ -123,11 +159,15 @@
 
 ## 0.2.43
 
+*Released: 2026-01-30T16:59:27Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.34.4
 
 ## 0.2.42
+
+*Released: 2026-01-29T14:02:02Z*
 
 ### Patch Changes
 
@@ -136,12 +176,16 @@
 
 ## 0.2.41
 
+*Released: 2026-01-16T13:57:39Z*
+
 ### Patch Changes
 
 - Updated dependencies [be67fc4]
     - @sap-ux/project-access@1.34.2
 
 ## 0.2.40
+
+*Released: 2026-01-15T14:14:39Z*
 
 ### Patch Changes
 
@@ -150,12 +194,16 @@
 
 ## 0.2.39
 
+*Released: 2026-01-14T17:56:49Z*
+
 ### Patch Changes
 
 - Updated dependencies [b132944]
     - @sap-ux/project-access@1.34.0
 
 ## 0.2.38
+
+*Released: 2026-01-12T12:01:34Z*
 
 ### Patch Changes
 
@@ -164,6 +212,8 @@
 
 ## 0.2.37
 
+*Released: 2026-01-09T11:35:48Z*
+
 ### Patch Changes
 
 - Updated dependencies [e111d0d]
@@ -171,12 +221,16 @@
 
 ## 0.2.36
 
+*Released: 2025-12-19T11:36:13Z*
+
 ### Patch Changes
 
 - Updated dependencies [c7f9a60]
     - @sap-ux/project-access@1.33.0
 
 ## 0.2.35
+
+*Released: 2025-12-18T21:05:02Z*
 
 ### Patch Changes
 
@@ -186,6 +240,8 @@
 
 ## 0.2.34
 
+*Released: 2025-12-15T10:50:50Z*
+
 ### Patch Changes
 
 - 4ecfbe2: Chore - upgrade eslint devDependenies, convert to flat config and fix lint issues.
@@ -194,12 +250,16 @@
 
 ## 0.2.33
 
+*Released: 2025-12-12T12:38:21Z*
+
 ### Patch Changes
 
 - Updated dependencies [7217d7d]
     - @sap-ux/project-access@1.32.15
 
 ## 0.2.32
+
+*Released: 2025-12-10T20:04:51Z*
 
 ### Patch Changes
 
@@ -208,6 +268,8 @@
 
 ## 0.2.31
 
+*Released: 2025-12-10T09:08:56Z*
+
 ### Patch Changes
 
 - Updated dependencies [985223a]
@@ -215,11 +277,15 @@
 
 ## 0.2.30
 
+*Released: 2025-12-08T17:56:48Z*
+
 ### Patch Changes
 
 - 037a430: fix high severity Sonar issues
 
 ## 0.2.29
+
+*Released: 2025-12-04T10:42:12Z*
 
 ### Patch Changes
 
@@ -228,6 +294,8 @@
 
 ## 0.2.28
 
+*Released: 2025-11-27T14:58:09Z*
+
 ### Patch Changes
 
 - Updated dependencies [07725fe]
@@ -235,11 +303,15 @@
 
 ## 0.2.27
 
+*Released: 2025-11-26T14:39:37Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.32.10
 
 ## 0.2.26
+
+*Released: 2025-11-24T12:45:14Z*
 
 ### Patch Changes
 
@@ -249,6 +321,8 @@
 
 ## 0.2.25
 
+*Released: 2025-11-05T06:53:42Z*
+
 ### Patch Changes
 
 - cfe9c13: Add deep link to package and changelog to README.md
@@ -257,12 +331,16 @@
 
 ## 0.2.24
 
+*Released: 2025-10-27T13:41:31Z*
+
 ### Patch Changes
 
 - Updated dependencies [d895232]
     - @sap-ux/project-access@1.32.7
 
 ## 0.2.23
+
+*Released: 2025-10-22T09:32:12Z*
 
 ### Patch Changes
 
@@ -271,6 +349,8 @@
 
 ## 0.2.22
 
+*Released: 2025-10-21T07:47:50Z*
+
 ### Patch Changes
 
 - Updated dependencies [c5d7915]
@@ -278,11 +358,15 @@
 
 ## 0.2.21
 
+*Released: 2025-10-10T12:22:06Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.32.4
 
 ## 0.2.20
+
+*Released: 2025-10-06T17:09:01Z*
 
 ### Patch Changes
 
@@ -292,17 +376,23 @@
 
 ## 0.2.19
 
+*Released: 2025-10-06T10:53:53Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.32.2
 
 ## 0.2.18
 
+*Released: 2025-09-19T16:36:41Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.32.1
 
 ## 0.2.17
+
+*Released: 2025-09-19T00:44:48Z*
 
 ### Patch Changes
 
@@ -311,6 +401,8 @@
 
 ## 0.2.16
 
+*Released: 2025-09-18T07:30:28Z*
+
 ### Patch Changes
 
 - Updated dependencies [c385a76]
@@ -318,11 +410,15 @@
 
 ## 0.2.15
 
+*Released: 2025-09-16T06:28:11Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.14
 
 ## 0.2.14
+
+*Released: 2025-08-28T13:37:07Z*
 
 ### Patch Changes
 
@@ -330,11 +426,15 @@
 
 ## 0.2.13
 
+*Released: 2025-08-14T14:36:13Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.12
 
 ## 0.2.12
+
+*Released: 2025-08-13T10:00:10Z*
 
 ### Patch Changes
 
@@ -342,11 +442,15 @@
 
 ## 0.2.11
 
+*Released: 2025-08-11T13:59:55Z*
+
 ### Patch Changes
 
 - 944af3e: Update all UI5 OPA test runner HTML files to use relative resource paths instead of absolute paths
 
 ## 0.2.10
+
+*Released: 2025-08-01T09:39:39Z*
 
 ### Patch Changes
 
@@ -354,11 +458,15 @@
 
 ## 0.2.9
 
+*Released: 2025-07-30T19:33:46Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.9
 
 ## 0.2.8
+
+*Released: 2025-07-28T22:57:52Z*
 
 ### Patch Changes
 
@@ -367,6 +475,8 @@
 
 ## 0.2.7
 
+*Released: 2025-07-08T12:06:04Z*
+
 ### Patch Changes
 
 - Updated dependencies [58cdce6]
@@ -374,11 +484,15 @@
 
 ## 0.2.6
 
+*Released: 2025-07-04T09:34:59Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.6
 
 ## 0.2.5
+
+*Released: 2025-07-03T12:14:55Z*
 
 ### Patch Changes
 
@@ -387,6 +501,8 @@
 
 ## 0.2.4
 
+*Released: 2025-06-27T16:39:38Z*
+
 ### Patch Changes
 
 - Updated dependencies [f75b89d]
@@ -394,11 +510,15 @@
 
 ## 0.2.3
 
+*Released: 2025-06-24T07:18:46Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.3
 
 ## 0.2.2
+
+*Released: 2025-05-30T07:18:36Z*
 
 ### Patch Changes
 
@@ -406,11 +526,15 @@
 
 ## 0.2.1
 
+*Released: 2025-05-15T13:36:04Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.30.1
 
 ## 0.2.0
+
+*Released: 2025-05-14T22:35:53Z*
 
 ### Minor Changes
 
@@ -423,6 +547,8 @@
 
 ## 0.1.21
 
+*Released: 2025-05-05T13:48:20Z*
+
 ### Patch Changes
 
 - Updated dependencies [ea0e2c0]
@@ -430,11 +556,15 @@
 
 ## 0.1.20
 
+*Released: 2025-04-24T15:01:01Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.21
 
 ## 0.1.19
+
+*Released: 2025-04-23T15:22:38Z*
 
 ### Patch Changes
 
@@ -443,17 +573,23 @@
 
 ## 0.1.18
 
+*Released: 2025-04-23T13:59:14Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.19
 
 ## 0.1.17
 
+*Released: 2025-04-09T12:10:50Z*
+
 ### Patch Changes
 
 - 3005a27: fix: make integrity check format free
 
 ## 0.1.16
+
+*Released: 2025-03-19T08:54:25Z*
 
 ### Patch Changes
 
@@ -462,11 +598,15 @@
 
 ## 0.1.15
 
+*Released: 2025-03-18T09:41:03Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.17
 
 ## 0.1.14
+
+*Released: 2025-03-10T20:40:35Z*
 
 ### Patch Changes
 
@@ -474,11 +614,15 @@
 
 ## 0.1.13
 
+*Released: 2025-03-07T09:29:37Z*
+
 ### Patch Changes
 
 - 5148043: Fix to ensure writing of encoded integrity data
 
 ## 0.1.12
+
+*Released: 2025-03-06T16:16:20Z*
 
 ### Patch Changes
 
@@ -487,6 +631,8 @@
 
 ## 0.1.11
 
+*Released: 2025-03-05T11:16:53Z*
+
 ### Patch Changes
 
 - Updated dependencies [3cc8f8a]
@@ -494,17 +640,23 @@
 
 ## 0.1.10
 
+*Released: 2025-03-05T09:14:02Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.13
 
 ## 0.1.9
 
+*Released: 2025-03-04T08:55:15Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.12
 
 ## 0.1.8
+
+*Released: 2025-02-28T15:10:48Z*
 
 ### Patch Changes
 
@@ -513,11 +665,15 @@
 
 ## 0.1.7
 
+*Released: 2025-02-27T14:15:03Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.10
 
 ## 0.1.6
+
+*Released: 2025-02-27T08:22:17Z*
 
 ### Patch Changes
 
@@ -526,12 +682,16 @@
 
 ## 0.1.5
 
+*Released: 2025-02-17T13:42:22Z*
+
 ### Patch Changes
 
 - Updated dependencies [c50e09f]
     - @sap-ux/project-access@1.29.8
 
 ## 0.1.4
+
+*Released: 2025-02-13T10:58:34Z*
 
 ### Patch Changes
 
@@ -540,17 +700,23 @@
 
 ## 0.1.3
 
+*Released: 2025-02-07T19:58:32Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.6
 
 ## 0.1.2
 
+*Released: 2025-02-05T17:08:20Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.5
 
 ## 0.1.1
+
+*Released: 2025-02-04T15:39:07Z*
 
 ### Patch Changes
 
@@ -559,11 +725,15 @@
 
 ## 0.1.0
 
+*Released: 2025-01-31T10:47:16Z*
+
 ### Minor Changes
 
 - 42fd8f6: Add function to check if Fiori project integrity is initialized
 
 ## 0.0.7
+
+*Released: 2025-01-30T10:04:50Z*
 
 ### Patch Changes
 
@@ -572,6 +742,8 @@
 
 ## 0.0.6
 
+*Released: 2025-01-28T09:20:06Z*
+
 ### Patch Changes
 
 - Updated dependencies [93ef8c1]
@@ -579,17 +751,23 @@
 
 ## 0.0.5
 
+*Released: 2025-01-27T18:57:16Z*
+
 ### Patch Changes
 
 - @sap-ux/project-access@1.29.1
 
 ## 0.0.4
 
+*Released: 2025-01-16T21:27:08Z*
+
 ### Patch Changes
 
 - cd2ee02: Cleanup test project
 
 ## 0.0.3
+
+*Released: 2025-01-16T02:22:29Z*
 
 ### Patch Changes