+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/project-integrity
 
+## 0.2.62
+
+### Patch Changes
+
+- Updated dependencies [f384ace]
+    - @sap-ux/project-access@1.35.15
+
+## 0.2.61
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.14
+
 ## 0.2.60
 
 ### Patch Changes